Use cast to avoid signed/unsigned warnings
git-svn-id: http://svn.python.org/projects/python/branches/alpha100@3753 6015fed2-1504-0410-9fe1-9d1591cc4771

+++ Modules/regexpr.c
@@ -644,7 +644,7 @@ regexp_t bufp;
 		  }
 		if (range)
 		  {
-		    for (a = prev; a <= ch; a++)
+		    for (a = prev; a <= (int)ch; a++)
 		      SETBIT(pattern, offset, a);
 		    prev = -1;
 		    range = 0;
